+++ src/os2datascanner/projects/report/organizations/admin.py
@@ -34,8 +34,22 @@ class AliasAdmin(ReadOnlyAdminMixin, admin.ModelAdmin):
 
 @admin.register(Account)
 class AccountAdmin(ReadOnlyAdminMixin, admin.ModelAdmin):
-    list_display = ('username', 'first_name', 'last_name', 'organization', 'manager', 'uuid')
-    readonly_fields = ('username', 'first_name', 'last_name', 'organization', 'manager', 'uuid')
+    list_display = (
+        'username',
+        'first_name',
+        'last_name',
+        'email',
+        'organization',
+        'manager',
+        'uuid',)
+    readonly_fields = (
+        'username',
+        'first_name',
+        'last_name',
+        'email',
+        'organization',
+        'manager',
+        'uuid',)
 
 
 @admin.register(AccountOutlookSetting)